Add semantics support to Markdown, move to right package
Usage in tests:

```kotlin
fun SemanticsNodeInteraction.assertMarkdownEquals(
  expected: String
): SemanticsNodeInteraction =
  assert(hasMarkdownExactly(expected))

private fun hasMarkdownExactly(expected: String): SemanticsMatcher =
  SemanticsMatcher.expectValue(RawMarkdown, expected)
```
